Moved class Param --> moco.ParamMoco
